Revised video call countdown to countupp
